Work in progress (#23)
* Update -mm-mount-mmojo-share.sh

* Update 304-Build-Cosmopolitan.md

* Update 201-Prepare-WSL.md

* Update 201-Prepare-WSL.md

* Update 203-Clone-Mmojo-Server-Repo.md

* Update 301-Re-Clone-Local-Mmojo-Server-Repo.md

* Update 401-Re-Clone-Mmojo-Server-Repo.md

* Update mm-update-local-mmojo-server-repo.sh

* Update mm-work-in-progress-branch.sh

* Update mm-main-branch.sh

* Update mm-work-in-progress-branch.sh

* Update mm-work-in-progress-branch.sh

* Update mm-main-branch.sh

* Update mm-update-local-mmojo-server-repo.sh

* Update mm-environment-variables.sh

* Update 404-Test-Debug-Command-Line.sh

* Update mm-update-local-mmojo-server-repo.sh

* Update mm-work-in-progress-branch.sh

* Update mm-main-branch.sh

* Update 404-Test-Debug-mmojo-server-args.sh

* Update 404-Test-Debug-mmojo-server-support.sh

* Update 405-Test-Cosmo-APE.sh

* Update 405-Test-Cosmo-aarch64.sh

* Update 405-Test-Cosmo-APE.sh

* Update 407-Test-CUDA.sh

* Update 406-Test-CPU.sh

* Update 405-Test-Cosmo-x86_64.sh

* Update 404-Test-Debug-Command-Line.sh

* Update 302-Download-Models.md

* Update 303-Copy-Models.md

* Update 304-Build-Cosmopolitan.md

* Update 305-Build-OpenSSL.md

* Update 306-Install-CUDA-Toolkit.md

* Update 307-Build-Vulkan-SDK.md

* Update 402-Prepare-to-Build.md

* Update 403-Prepare-to-Test.md

* Update 404-Build-Debug.md

* Update 405-Build-APE.md

* Update 406-Build-CPU.md

* Update 407-Build-CUDA.md

* Update 408-Build-Vulkan.md

* Update 501-Package-APE.md

* Update 502-Package-zip.md

* Update 503-Package-SEAS.md

* Update 407-Build-CUDA.md

* Update 400-Build-Mmojo-Server.md

* Update -Start-Here-TOC.md

* Delete instructions/401-Re-Clone-Mmojo-Server-Repo.md

* Update and rename 402-Prepare-to-Build.md to 401-Prepare-to-Build.md

* Update and rename 403-Prepare-to-Test.md to 402-Prepare-to-Test.md

* Update and rename 404-Build-Debug.md to 403-Build-Debug.md

* Update and rename 405-Build-APE.md to 404-Build-APE.md

* Update and rename 406-Build-CPU.md to 405-Build-CPU.md

* Update and rename 407-Build-CUDA.md to 406-Build-CUDA.md

* Update and rename 408-Build-Vulkan.md to 407-Build-Vulkan.md

* Update and rename 402-Clone-Repos.sh to 401-Clone-Repos.sh

* Update and rename 402-Customize-webui.sh to 401-Customize-webui.sh

* Update and rename 402-Fix-llama-cpp.sh to 401-Fix-llama-cpp.sh

* Update and rename 402-Rollback-webui.sh to 401-Rollback-webui.sh

* Update and rename 404-Build-Debug.sh to 403-Build-Debug.sh

* Update and rename 404-Test-Debug-Command-Line.sh to 403-Test-Debug-Command-Line.sh

* Update and rename 404-Test-Debug-mmojo-server-args.sh to 403-Test-Debug-mmojo-server-args.sh

* Update and rename 404-Test-Debug-mmojo-server-support.sh to 403-Test-Debug-mmojo-server-support.sh

* Update and rename 405-Assemble-Cosmo-APE-Local.sh to 404-Assemble-Cosmo-APE-Local.sh

* Update and rename 405-Assemble-Cosmo-APE-Mmojo-Share.sh to 404-Assemble-Cosmo-APE-Mmojo-Share.sh

* Update and rename 405-Build-Cosmo-aarch64.sh to 404-Build-Cosmo-aarch64.sh

* Update and rename 405-Build-Cosmo-x86_64.sh to 404-Build-Cosmo-x86_64.sh

* Update and rename 405-Copy-Cosmo-APE-to-Mmojo-Share.sh to 404-Copy-Cosmo-APE-to-Mmojo-Share.sh

* Update and rename 405-Copy-Cosmo-aarch64-to-Mmojo-Share.sh to 404-Copy-Cosmo-aarch64-to-Mmojo-Share.sh

* Update and rename 405-Copy-Cosmo-x86_64-to-Mmojo-Share.sh to 404-Copy-Cosmo-x86_64-to-Mmojo-Share.sh

* Update and rename 405-Test-Cosmo-APE.sh to 404-Test-Cosmo-APE.sh

* Update and rename 405-Test-Cosmo-aarch64.sh to 404-Test-Cosmo-aarch64.sh

* Update and rename 405-Test-Cosmo-x86_64.sh to 404-Test-Cosmo-x86_64.sh

* Update and rename 406-Build-CPU.sh to 405-Build-CPU.sh

* Update and rename 406-Copy-CPU-Build-to-Mmojo-Share.sh to 405-Copy-CPU-Build-to-Mmojo-Share.sh

* Update and rename 406-Test-CPU.sh to 405-Test-CPU.sh

* Update and rename 407-Build-CUDA.sh to 406-Build-CUDA.sh

* Update and rename 407-Copy-CUDA-Build-to-Mmojo-Share.sh to 406-Copy-CUDA-Build-to-Mmojo-Share.sh

* Update and rename 407-Test-CUDA.sh to 406-Test-CUDA.sh

* Update and rename 408-Build-Vulkan.sh to 407-Build-Vulkan.sh

* Update and rename 408-Copy-Vulkan-Build-to-Mmojo-Share.sh to 407-Copy-Vulkan-Build-to-Mmojo-Share.sh

* Update and rename 408-Test-Vulkan.sh to 407-Test-Vulkan.sh

* Update 401-Clone-Repos.sh

* Update 401-Customize-webui.sh

* Update 401-Fix-llama-cpp.sh

* Update 401-Rollback-webui.sh

* Update 403-Build-Debug.sh

* Update 403-Test-Debug-Command-Line.sh

* Update 403-Test-Debug-mmojo-server-args.sh

* Update 403-Test-Debug-mmojo-server-support.sh

* Update 404-Assemble-Cosmo-APE-Local.sh

* Update 404-Assemble-Cosmo-APE-Mmojo-Share.sh

* Update 404-Build-Cosmo-aarch64.sh

* Update 404-Build-Cosmo-x86_64.sh

* Update 404-Copy-Cosmo-APE-to-Mmojo-Share.sh

* Update 404-Copy-Cosmo-aarch64-to-Mmojo-Share.sh

* Update 404-Copy-Cosmo-x86_64-to-Mmojo-Share.sh

* Update 404-Test-Cosmo-APE.sh

* Update 404-Test-Cosmo-aarch64.sh

* Update 404-Test-Cosmo-x86_64.sh

* Update 405-Build-CPU.sh

* Update 405-Copy-CPU-Build-to-Mmojo-Share.sh

* Update 405-Test-CPU.sh

* Update 406-Build-CUDA.sh

* Update 406-Copy-CUDA-Build-to-Mmojo-Share.sh

* Update 406-Test-CUDA.sh

* Update 407-Build-Vulkan.sh

+++ scripts/mm-update-local-mmojo-server-repo.sh
@@ -7,23 +7,25 @@
 # See licensing note at end.
 ################################################################################
 
-WD=$(pwd)
-cd $MMOJO_SERVER_DIR
-git reset --hard
-git pull
-cd $WD
+if [ -d "$MMOJO_SERVER_DIR" ]; then
+  WD=$(pwd)
+  cd $MMOJO_SERVER_DIR
+  git reset --hard
+  git pull
+  cd $WD
 
-chmod a+x $MMOJO_SERVER_SCRIPTS/2*.sh
-chmod a+x $MMOJO_SERVER_SCRIPTS/3*.sh
-chmod a+x $MMOJO_SERVER_SCRIPTS/4*.sh
-# chmod a+x $MMOJO_SERVER_SCRIPTS/5*.sh
+  # These are the scripts. They need to be executable.
+  chmod a+x $MMOJO_SERVER_SCRIPTS/2*.sh
+  chmod a+x $MMOJO_SERVER_SCRIPTS/3*.sh
+  chmod a+x $MMOJO_SERVER_SCRIPTS/4*.sh
+  # chmod a+x $MMOJO_SERVER_SCRIPTS/5*.sh
 
-### Links don't work - end up modifying repo files on chmod.
-# for FILE in $MMOJO_SERVER_SCRIPTS/mm-*.sh; do
-#  ln -sf $FILE $HOME_SCRIPTS
-# done
-cp $MMOJO_SERVER_SCRIPTS/mm-*.sh $HOME_SCRIPTS
-chmod a+x $HOME_SCRIPTS/mm-*.sh
+  ### Links don't work - end up modifying repo files on chmod.
+  cp $MMOJO_SERVER_SCRIPTS/mm-*.sh $HOME_SCRIPTS
+  chmod a+x $HOME_SCRIPTS/mm-*.sh
+else
+  echo "The $MMOJO_SERVER_DIR directory does not exist."
+fi
 
 printf "\n**********\n*\n* FINISHED: mm-update-local-mmojo-server-repo.sh.\n*\n**********\n\n"
